Fix for uncaught promise exceptions in libsignal.
Libsignal had a queue mech that would dangle a promise off a users
queued work for the purposes of cleanup/gc of their internal queue
data structures.  This dangling promise was only hooked on the
onResolved callback which had the effect of exposing all exceptions
(caught and uncaught) as being programming bugs, when they were
in fact still propogated up the stack in the previous link in the
promise chain.  Frankly, I don't like the way they chain promises as a
form of queuing.  It should be an actual stack of jobs instead to avoid
other stack polution issues in addition to this wart.