Functional UX for message flipping.
Added little buttons for fliping message back and forth between the
normal and moreinfo sides.

There is a bug (possibly a browser one) that makes the animation abort
if the cursor is moved outside the message box during the animation but
it still lands on the right side of things.  Hopefully not too annoying
that we can't use the feature.