Delivery receipt upgrades, textsecure cleanups, list-view checkpoint
Checkpoint work for revamped delivery recipts (can track who actually
ack'd our messages in the message object now.).

Revamping the conversation collection and how it ties to it's views
along with how ListView works.  It has some terrible bugs, some of which
are related to backwards sorting of the ConversationCollection (fixed)
now.