Refactor websocket resource (fix connect races).
The way websocket resource managed websockets led to race conditions on
startup. It needs to internally manage the websocket and also handle
event listener attachment to that websocket.

Extracted the connect phase of the websocket to a public call so
websocket receiver users like message-receiver can do all their setup
and then ask for the network to come online.   This kills all the races.

Fixes #92