Remove safari-mobile incompat welcome elements.
Some elements in the welcome screen didn't render properly in safari
mobile.  The language is also a bit misleading as it talks about
entering test for help but there is no place for the user to enter text
on that screen.  I think we should revisit this with some interactive
compontents for directing them to the correct conversation starting and
help starting buttons.