Liberate provisioning in dev / Remove superflous version query on sw.
Let provisioning take place on dev stacks that don't have superman's
number explicitly coded in.  Otherwise they would bail since they
didn't know who that address was.

Service worker doesn't need a version query arg on it.  It's setup on
the server side to never cache and it leads to proliferation of service
worker registries when the URL changes.