Redesign ListView  - Fixes numerous race conditions.
The state change queueing for ListView was doing some changes outside
the queue which led to corruption.  In addition to that, the code no
longer makes assumptions about the DOM stability and will directly use
the DOM elements for positioning changes.  This is faster and safer,
given that the nodes could in theory be manipulated by external forces.

Fixes several DnD and pin positioning issues.