Fix issue where thread only shows new messages.
If you get messages for a never opened thread and then switch to it, we
would wrongly optimize away the database fetch because the number of
messages in the thread was already non-zero.  We need to instead check
to see if there is less than the page size or total message count.