Working identity key system.
This is a working draft.  There is much to be said and changed about the
UX.  Namely need for more explanation and hand holding.   It's complex
and hard to understand at present.

Messages are released from quarantine any time a user is trusted or
untrusted.   Only messages received while the current trust is different
from the current identity key are stored in quarantine.  There is no real
option for complete purge of the user and there messages at present.  So
you can't block/flag a user yet.