Include package version directly into app/sw bundles.
In some cases the version is reported incorrectly because /@env.js is
loaded separately from the application javascript.  The caching
behavior is distinct and thus a user can end up believing they are on a
newer version of code than they are.  This is especially prevalent in
service workers that rarely update.