Major HTML injection fix for markdown formatting.
I was setting html content during markdown processing.  This is very
wrong given that the source content was a #text node.  So now we set
content using Element.textContent which is safe way to avoid creation of
new DOM that was never intended to be parsed.  Namely, this fixes
literals that had html entities in them, such as `Hello <Bob>`.