POC edge support via indexeddb polyfill.
This integrates the unmaintained and unsupported ie-gap polyfill library
made to support compound and multiEntry indexes in Microsoft's IE
browsers.  It's been modified to work with Edge instead and various
portions have been modified to either fix it or adapt it to our target
ES level.  Basic functionality does work in limited testing, but it's
still POC quality.