Fix for store.loadSession(encodedAddr)
Sending to a specific device was corrupting the session cache because we
would only load one entry.  A cache hit later would then only have one
entry instead of all the sessions for the address in question.  This
commit modifies the cache miss handler to load all addresses which
prevents any mishaps later.