Refactor ListView for concurrent rendering.
This has pros and cons.  The pros are that we don't need async
serialization for the add/remove/reset methods and the rendering runs
faster.  The con is that DOM layout happens more sporadically for reset,
since it's not done after rendering completes for each item (subject to
change however).

For addModel performance this is vastly faster though.