Handle embed client in same browser as full client.
Tune the isolation params for duplicate session detection to exclude
sessions on the same site acting with different users.  E.g. if you are
logged in as User A in one tab, then user B is free to work work in
another tab.  This normally wouldn't be possible except in the case of
using /@embed clients which have ephemeral users along side a full
client.