Fix calling bug which prevented media track stop.
Prevent call view from recreating a media stream during a leave()
operation triggered by a view removal.  E.g. We are in tear down so avoid
needlessly getting the media stream again.

Move the media stream track stop calls to the final stage of calling
cleanup so that regardless of what happens during leave we stop
the video and audio tracks.  This lets the browser and OS turn off the
audio and video hardware (and the associated indicators).

Also switched to F.log for logging.

Fixes #277