Implement callHeartbeat control.
callHeartbeat is a thread wide gossip control message that indicates to
thread members of an ongoing call.

This client uses the heartbeat to keep the thread callActive timestamp
updated.  This timestamp is used to indicate if anyone is on a call in
a thread.

Added callactivity status to the nav-item view so you can see which
threads are involved in a call (even if you are not personally attending
the call).

Fixes #273