Allow caches to work without indexeddb ready state.
For caches based on indexeddb, allow them to work in a memory-only state
while indexeddb is not available.  This is required for some resources
to be able to function when indexeddb is not available.  Such as loading
templates to display information to a user about compatibility issues.

This also allows the cache to start working before indexeddb is fully
ready and as soon as it is, the data will be flushed to idb for future
loads.